refactored request and response packages
