+++ src/main/java/com/pratap/blog/controller/PostController.java
@@ -3,9 +3,9 @@ package com.pratap.blog.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.pratap.blog.constant.AppConstant;
 import com.pratap.blog.dto.PostDto;
-import com.pratap.blog.model.PostPageableResponseModel;
-import com.pratap.blog.model.PostRequestModel;
-import com.pratap.blog.model.PostResponseModel;
+import com.pratap.blog.model.response.PostPageableResponseModel;
+import com.pratap.blog.model.request.PostRequestModel;
+import com.pratap.blog.model.response.PostResponseModel;
 import com.pratap.blog.service.PostService;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
