+++ src/main/java/com/pratap/blog/model/response/JWTAuthResponse.java
@@ -1,4 +1,4 @@
-package com.pratap.blog.model;
+package com.pratap.blog.model.response;
 
 import lombok.Getter;
 import lombok.Setter;
