+++ src/main/java/com/pratap/blog/service/impl/PostServiceImpl.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.pratap.blog.dto.PostDto;
 import com.pratap.blog.entity.Post;
 import com.pratap.blog.exception.ResourceNotFoundException;
-import com.pratap.blog.model.PostPageableResponseModel;
-import com.pratap.blog.model.PostResponseModel;
+import com.pratap.blog.model.response.PostPageableResponseModel;
+import com.pratap.blog.model.response.PostResponseModel;
 import com.pratap.blog.repository.PostRepository;
 import com.pratap.blog.service.PostService;
 import com.pratap.blog.utils.IdGeneratorUtil;
